+++ src/fluid/operators/advection.hpp
@@ -1,26 +1,28 @@
-#include "common.hpp"
+#include "../common.hpp"
 
-inline double limit(double x, double min, double max);
+inline float limit(float x, float min, float max);
 
 template <size_t NX, size_t NY>
-void advect_linear_backtrace(const FluidArray<NX, NY>& input, FluidArray<NX, NY>& output, const FluidArray<NX, NY>& xvel, const FluidArray<NX, NY>& yvel, double dx, double dt) {
+void advect_linear_backtrace(const FluidArray<NX, NY>& input, FluidArray<NX, NY>& output, const FluidArray<NX, NY>& xvel, const FluidArray<NX, NY>& yvel, float dx, float dt) {
 
-    double dt0 = dt / dx;
-    for (int i = 1; i < NX - 1; ++i) {
-        for (int j = 1; j < NY - 1; ++j) {
-            double x = i - dt0 * xvel(i, j);
-            double y = j - dt0 * yvel(i, j);
+    float dt0 = dt / dx;
+    for (size_t i = 1; i < NX - 1; ++i) {
+        for (size_t j = 1; j < NY - 1; ++j) {
+            float x = i - dt0 * xvel(i, j);
+            float y = j - dt0 * yvel(i, j);
 
-            x = limit(x, 0.5, NX - 1.5);
-            y = limit(y, 0.5, NY - 1.5);
+            x = limit(x, 0.5f, NX - 1.5f);
+            y = limit(y, 0.5f, NY - 1.5f);
 
-            int i0 = (int)x; int i1 = i0 + 1;
-            int j0 = (int)y; int j1 = j0 + 1;
+            unsigned int i0 = (unsigned int)x;
+            unsigned int i1 = i0 + 1;
+            unsigned int j0 = (unsigned int)y;
+            unsigned int j1 = j0 + 1;
 
-            double x1 = x - i0;
-            double x0 = 1 - x1;
-            double y1 = y - j0;
-            double y0 = 1 - y1;
+            float x1 = x - i0;
+            float x0 = 1 - x1;
+            float y1 = y - j0;
+            float y0 = 1 - y1;
 
             output(i, j) = x0 * (y0 * input(i0, j0) + y1 * input(i0, j1)) +
                            x1 * (y0 * input(i1, j0) + y1 * input(i1, j1));
@@ -29,7 +31,7 @@ void advect_linear_backtrace(const FluidArray<NX, NY>& input, FluidArray<NX, NY>
     }
 }
 
-inline double limit(double x, double min, double max) {
+inline float limit(float x, float min, float max) {
     if (x < min) {
         x = min;
     } else if (x > max) {
